Fix f0 interp and cyclic imports, rename classes
