+++ configs/acoustic_v3.yaml
@@ -1,4 +1,4 @@
-bases: ["resources/base_v3.yaml"]
+bases: ["configs/base_v3.yaml"]
 
 data:
   dictionaries:
